+++ packages/person/src/index.ts
@@ -1,0 +1,3 @@
+export * from './person-avatar';
+export * from './person-provider';
+export * from './types';