+++ packages/person/src/person-avatar/types.ts
@@ -1,0 +1,8 @@
+import { PersonElementProps } from '../person';
+import { AvatarSize } from '@equinor/fusion-wc-avatar';
+
+export type PersonAvatarElementProps = PersonElementProps & {
+  size?: AvatarSize;
+  clickable?: boolean;
+  disabled?: boolean;
+};