+++ packages/chip/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.6](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-chip@0.2.5...@equinor/fusion-wc-chip@0.2.6) (2021-10-29)
+
+**Note:** Version bump only for package @equinor/fusion-wc-chip
+
+
+
+
+
 ## [0.2.5](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-chip@0.2.4...@equinor/fusion-wc-chip@0.2.5) (2021-10-28)