+++ packages/skeleton/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.8](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-skeleton@0.2.7...@equinor/fusion-wc-skeleton@0.2.8) (2021-10-29)
+
+**Note:** Version bump only for package @equinor/fusion-wc-skeleton
+
+
+
+
+
 ## [0.2.7](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-skeleton@0.2.6...@equinor/fusion-wc-skeleton@0.2.7) (2021-10-28)