+++ storybook/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.15](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-storybook@1.14.14...@equinor/fusion-wc-storybook@1.14.15) (2021-10-29)
+
+**Note:** Version bump only for package @equinor/fusion-wc-storybook
+
+
+
+
+
 ## [1.14.14](https://github.com/equinor/fusion-web-components/compare/@equinor/fusion-wc-storybook@1.14.13...@equinor/fusion-wc-storybook@1.14.14) (2021-10-28)